Standardise PDF renderer
Extract a common PDF renderer component and reduce the amount of repeated rendering code used across page routers.

Adds a `noRender` option to all PDFs. This will prevent the PDF renderer assuming we're passing a mustache template and trying to check over a several megabyte string for a mustache token regex.
